Print clearer error if parsing fails
This implementation is not perfect as I feel parse_roll_string should
not return an Option but rather a Result. However, it is not trivial to
propagate the Result's we get from functions called internally because
they do not have the same type. I'll probably figure out a way to clean
this up, but for now it works.