Small `mut` cleanup
`roll_substrings` is only `mut` because of its multiline initialization.
With this commit it's clear where initialization ends and
`roll_substrings` "becomes const".

Consider refactoring it into a separate function in the future. It may
sound unnecessary because it's short, but I think it will improve
readability. For example, calling a function `split_roll_string` is a
cleaner communication of intent than what we currently do.